pvpg_parallel and pvpg_concise now have the option of showing the fraction of segments that have canopy cover photons. This could be turned into a flag by throwing out tracks under some fraction, which I'm sure will be investigated later. Removed functions pvpg, pvpg_flagged_penalized, and plot_pvpg because they are unilateral worse (and not as cool) and pvpg_concise. Just taking up scroll space. Next step is to transition to actual data by a snow camera.
